refactor(post.go): add the viper support for the command
