+++ core/packages/delete.go
