+++ core/packages/put.go
@@ -11,7 +11,7 @@ import (
 	"github.com/go-resty/resty/v2"
 )
 
-func Create() (getOne, error) {
+func Update(id string) (getOne, error) {
 	var dataStruct getOne
 	name := terminal.BasicPrompt("Name", "")
 	author := terminal.BasicPrompt("Author", "")
@@ -19,17 +19,18 @@ func Create() (getOne, error) {
 	description := terminal.BasicPrompt("Description", "")
 	version := terminal.BasicPrompt("Version", "")
 	password := terminal.PasswordPrompt("Password")
+	newPassword := terminal.PasswordPrompt("New password")
 
-	if name == "" || author == "" || url == "" || description == "" || version == "" || password == "" {
+	if name == "" || author == "" || url == "" || description == "" || version == "" || password == "" || newPassword == "" {
 		colors.Warn("Fill all blanks")
-		return dataStruct, errors.New("Fill all blanks")
+		return dataStruct, errors.New("fill all blanks")
 	}
 
 	client := resty.New()
 	resp, err := client.R().
 		SetHeader("Content-Type", "application/json").
-		SetBody(fmt.Sprintf(`{"name":"%v", "author":"%v", "url":"%v", "description":"%v", "version":"%v", "password":"%v"}`, name, author, url, description, version, password)).
-		Post("https://moldy-api.herokuapp.com/api/v1/packages/new")
+		SetBody(fmt.Sprintf(`{"name":"%v", "author":"%v", "url":"%v", "description":"%v", "version":"%v", "password":"%v", "newpassword":"%v"}`, name, author, url, description, version, password, newPassword)).
+		Put("https://moldy-api.herokuapp.com/api/v1/packages/update/" + id)
 
 	functions.CheckErrors(err, "Code 5", "The fetch to the API failed", "Unknown solution, API Internal server error")
 	if err := json.Unmarshal(resp.Body(), &dataStruct); err != nil {
@@ -40,8 +41,8 @@ func Create() (getOne, error) {
 		colors.Error(dataStruct.Message)
 		return dataStruct, errors.New(dataStruct.Message)
 	}
-	colors.Success("\nCreated successfully")
-	fmt.Printf("\nThis is the data of the package created:\n\nID: %v\nName: %v\nAuthor: %v\nDescription: %v\nURL: %v\nVersion: %v\n", dataStruct.Data.Id, dataStruct.Data.Name, dataStruct.Data.Author, dataStruct.Data.Description, dataStruct.Data.Url, dataStruct.Data.Version)
-	fmt.Println("\nIf some of this data not is correct you can update it using\n\nmoldy update <id>\n\nGo to the docs to know more about this command")
+
+	colors.Success("Updated successfully")
+	fmt.Printf("\nThis is the data of the package now:\nID: %v\nName: %v\nAuthor: %v\nDescription: %v\nURL: %v\nVersion: %v\n", dataStruct.Data.Id, dataStruct.Data.Name, dataStruct.Data.Author, dataStruct.Data.Description, dataStruct.Data.Url, dataStruct.Data.Version)
 	return dataStruct, nil
 }
